api: Throw an error on invalid JSON in response.
The Zulip API guarantees that the response will always be JSON,
whether for an error or a success response.

If it's not, we should definitely treat that as an error, regardless
of the HTTP status code.  Because there is no JSON that parses to
`undefined` in JavaScript, the `undefined` that this `catch` handler
returns serves as a nice sentinel.

This can never happen anyway if the Zulip server is properly
configured and there isn't some bug in Zulip (and I'm not aware of any
bug we've had in the server code that would cause this particular
failure mode.)  But the Internet is big, and there are servers out
there that are misconfigured, so we should handle it properly.

In the wild, we've seen this (#2858) in a Sentry crash report, where
a request to /server_settings on a particular server produced an HTML
login page.  Moreover, that response was with HTTP status 200, so our
API layer would cheerfully return `undefined` to the caller, producing
a crash later when some other code innocently tried to look up a
property on the resulting `serverSettings` value.  With this fix,
trying to visit the same server from RealmScreen now produces an
appropriate error message, "Cannot connect to server", nicely
presented on that screen.

Fixes #2858.