android: Remove slash from a dependency's Gradle name.
Newer versions of Gradle don't allow project names to contain a
variety of characters, including `/`, that have special meaning in
filesystems or in various Gradle features:
  https://docs.gradle.org/4.2/release-notes.html#path-separator-characters-in-names-are-deprecated

We don't use such a new version of Gradle for the actual build, but
apparently the current versions of both Android Studio, and the
default Java extension in VS Code, do use it.  Both give obscure
errors (though AS's is much more reasonable) with the old name, and
successfully complete a Gradle project sync after this fix.

The project name is just an identifier for our own use within the
Gradle config, so we can rename it however we like; just need to do
so everywhere in our Gradle config.

An underscore is the choice made by newer upstream versions of
`react-native link`: https://github.com/facebook/react-native/pull/18275
so an underscore it is.