android: Share a bit of Android Studio configuration.
I just spent a few minutes poking around the Android Studio settings
to get things set up more nicely (and get rid of the warning popup
that the project was under a Git working copy it didn't know about).
It'd be a shame for everyone to have to duplicate that, and everyone
end up with a different subset of useful settings; better for us all
to share and for someone newly picking up the project to get the best
settings we can give them from the outset.

Upstream actually recommends keeping the entire .idea/ directory in
version control, except for the one file workspace.xml (which has
things like which windows you have open and where on your screen):
  https://www.jetbrains.com/help/idea/2017.1/about-projects.html

I'm not yet convinced that makes sense for everything, but it
certainly seems to for this file; so let's start here.