webview: Add `appendAuthToImages` inside webview's JS.
This is a better alternative to the function named the same
found in `url.js`.

There are numerous advantages in implementing this inside the
webview vs outside in the React Native world. First, we have a
pefect HTML parser - the browser. Instead of trying to devise a
smart RegEx that will not fail (tricky) or import and use an
HTML parser (not performant) we get that for free.

The function goes through all `img` tags in the page. Importantly,
it skips any image that is not in our realm. If it did pass the
`apiKey` to every image it would leak the secret key!

Very conveniently, the `src` property in the DOM is always the
full URL -- including the domain -- even if the `<img src="...">`
tag in the HTML source provided a relative URL.

The old images does not include any URL parameters, so we need to
use `?` to append ours; but the new thumbnails do already contain
one, so take this into account and use '&' instead.