docs: Move `push-notifications` into howto directory.
I was looking for this just now, and was puzzled not to see it (wasn't
there a PR with a writeup about this?  did I fail to merge it?  let's
look back through open PRs...).  I actually found it from a search of
the PRs, but the PR had indeed been merged -- it just put the doc
out here. :-)  Move it to be next to our other similar docs.