webview render: Refer to state.flags instead of message.flags.
This is part of fixing #2676.  As discussed in the PR thread #2843,
the `flags` property on a message object in our Redux store never
changes after we first receive the message, and instead we maintain an
updated version under the separate top-level state subtree `flags`;
but we've been using the message object's stale `flags` for rendering
the message in the message list.

This change switches to using the up-to-date flags from `state.flags`,
via the `getFlags` selector, instead.  Now when the user stars or
unstars a message, we still don't yet update the message list (that's
for a followup commit); but now at least the next message list we
render from scratch will reflect the update.

In making this change, we get a bit of a head start because we already
invoke `getFlags` to populate the `flags` prop of `MessageList` (in
its `connect` call), which gets passed down to `renderMessagesAsHtml`.
So we just need to use it there, and adapt the layers underneath to
take the different data format found in `state.flags`.

Of the two spots where we actually use `flags` in message rendering:

 * `messageTagsAsHtml` gets a bit simpler; it only needs to know
   whether or not the message is starred, so we just pass it a boolean
   for that.

 * `messageDiv` actually uses the flags as a list of strings (for now)
   to generate HTML data-* attributes, so we make a little helper
   function for it to just compute that list.

[greg: explained more context in commit message]