api: Fix noisy console.error on API error responses.
This code was broken in two ways:

* It caused a red-screen on any API error response, even a perfectly
  normal one like an expired server event queue.

* The bulk of the resulting error message consisted of internals of
  the `response` object that are meaningless when debugging
  application code, and conversely the actual error data from the
  server was not displayed.

Fix the first issue by downgrading to `console.warn`, and the second
by replacing `response` with the same meaningful data we include in
the error object we're about to throw.

This does give up a feature that in the right circumstances was
helpful for debugging: now when an API request fails in debug mode,
it's flagged in the console (so `adb logcat` and the Xcode equivalent)
but not on the app's screen.  To make the latter happen again, it'll
be enough to delete our `disableYellowBox` line so we can use RN's
"yellow-box" feature again.