docs/arch: Add a "React, etc." design doc.
Inspired in part by the recent d93cf766d, and also by discussions
where a doc like this might have been handy to refer to.  Also, this
is a bunch of stuff that wasn't obvious to *me* as I was getting to
know the codebase; hopefully this helps future folks come up to speed
faster.

I'm sure there are ways this can be improved; feedback and/or PRs
welcome.  In particular there are probably some errors; there are
definitely places where a useful link to upstream docs (for React or
Redux or something else) could be added; and there are surely areas
that are unclear.