messages: Fix and re-enable most tests after normalizing `narrows`.
The recent "messages: Normalize `narrows` state`" requires a bunch of
tests to be updated that contain fragments of our message state.  To
keep that commit's diff readable, we just disabled them there; we need
to fix and re-enable them.

This commit does that for most of the tests, where there's nothing
interesting to say about them.