messages: Fix and re-enable flags reducer tests.
The recent "messages: Normalize `narrows` state`" requires a bunch of
tests to be updated that contain fragments of our message state.  To
keep that commit's diff readable, we just disabled them there; we need
to fix and re-enable them.

This looks like we're doing something wrong, because we're fiddling
with an server event action, when we should only update state objects.
However, the `allMessages` part of this action is taken from the
state when the action is constructed.