flags: Update rehydrate handler for `messages` -> `narrows` rename.
This had been masked by a $FlowFixMe. :'-(  So, also fix that to
reflect good "fixme" practices:
 * The fixme is on as narrow a piece of code as possible; here,
   on just the `Object.values` call, not constructing its argument.
 * The fixme identifies the issue that makes a fixme necessary;
   here, the sloppy type `flow-types` gives to `Object.values`.

If the fixme had been on this narrow piece of code in the first place,
the type-checker would have spotted the need for the rename to be
applied here when the rename originally happened, in 638cddce0.