topic list: Drop "onBack" hack which prevented renders.
Our `connectPreserveOnBackOption` is a hack that stops Redux from
causing a component to re-render in certain cases while we're
performing a navigation.  See details in its jsdoc.

It's not a good solution anywhere.  Here, it causes a conspicuous
bug: #2808.

Moreover, the situation it's meant to mitigate is one we already fixed
here the right way a while ago, in adc07349f and its parent as part of
from its props, and pays no attention to the nav state.

So, just cut the hack out.

Fixes #2808.