accountScreen: Drop "onBack" hack which prevented renders.
Much like the previous commit; see also the jsdoc on
`connectPreserveOnBackOption` for details on what motivated this hack
and how it's broken.

Unlike the previous commit, we don't know of a specific user-visible
bug this use of the hack is causing right now.  But, as in the
previous commit, it doesn't do anything helpful here; this is another
place where we fixed the underlying problem the right way in #2746.

To confirm there isn't anything for the hack to help with here:

(Best is to test on slow old phones, you will able to spot any extra
re-render/side effect).

Or put a console.log inside `render` and check how many times it is
called. Thus all calls of re-render are required, or some calls can be
ignored. Pay attension when screen pops out it should not re-render.