ModalSearchNavBar: Use `getCanGoBack` rather than whole `getNav`.
We don't need to use the `getNav` selector in `ModalSearchNavBar`
as we use it only to check if we can go back, which is the purpose
of the `getCanGoBack` selector.

This is good for maintainability but also helps with a potential
future removal of the Redux-React Navigation integration.  This
is nearly the last component that directly refers to the whole
`nav` state.

The other is `AppWithNavigation`, and it serves only for that
Redux integration itself.