messages: Initialize new state subtrees in migrations.
Each migration should take a state that was a valid value of the old
`GlobalState` type, and return a state that is a valid value of the
new `GlobalState` type.

In particular, when we create a new state subtree (i.e. a new property
on `GlobalState`), the migration should initialize that subtree.

The reason we've largely gotten away without doing this (we've had no
migrations at all until recently, so certainly haven't been doing
this) is that only a small amount of code is exposed to any failure
here -- only REHYDRATE handlers, when handling the `payload` of the
action -- because the actual Redux state is always initialized in the
standard Redux way, by invoking the handlers with `undefined` previous
state.  In cases where this is complicated to do, we might choose to
skip it in the future too.

But here we have an actual, critical, bug that this prevents.  When
starting up the app, with a state left behind by a version from before
these migrations, we never get past the green loading screen; and the
log shows `TypeError: Object.values called on non-object` inside a
function named `rehydrate`, which must be the one in `flagsReducers`.
That's because `action.payload` has no property `messages` -- contrary
to our type annotation saying a REHYDRATE action's payload should be a
`GlobalState`.

And fixing it once and for all in these migrations (rather than worry
about future bugs if future changes we make to rehydrate code
innocently make this same assumption) is easy.  So do that.