messageReducers: Store forgotten fields on a reaction event.
We had a type error, which Flow awkwardly does not find.  (This is
facebook/flow#6871 .)  As a result, we were dropping these two
properties in our Redux store, which ends with us failing to include
them when we ask the server to add or remove our own reaction after
the user taps the reaction's display in the UI.

Fixes #2784.

[greg: added discussion of type error]