lightbox: Update dependency to fix breakage on iOS (#2654).
Fixes #2654.  The cause of that issue was that the PhotoView
component, from the library `react-native-photo-view`, was ignoring
headers passed in the `source` prop; without the `Authorization`
header, the Zulip server refuses access.

This was reported upstream:
  https://github.com/alwx/react-native-photo-view/issues/101
and was fixed in master with commit c58fd6b30:
  https://github.com/alwx/react-native-photo-view/pull/127
but the fix has yet to be published on NPM.

So, refer in our package.json directly to that commit, using the
special abbreviated syntax for a GitHub URL.  For reference, see:
  https://docs.npmjs.com/files/package.json#github-urls