autocomplete: Avoid rendering a huge list of emoji choices.
Fixes #2987.  When you type a message with a colon in it, like
"Here you go:", and try to send it... the app effectively freezes.

On closer study, everything gets unstuck if you wait a minute or two.
And one thing that's certainly happening in this case is that our
`AutocompleteView` is making an `EmojiAutocomplete` with a filter of
just ''; the resulting list of emoji will be all ~1500 or so of them.
The `AutocompleteView` ensures the `EmojiAutocomplete` isn't *visible*
when the filter is empty... but it is *there*.

We could kludge this by short-circuiting things when the filter is
empty, but we'd still have a lot of emoji after typing something
that's a popular prefix for emoji names... like `:s`, which matches a
total of 222 emoji in our standard set.  And, in fact, trying the
repro with a message like ":s" produces the same symptoms -- just for
not for quite as long.

On the other hand, trying this with `@` or `#`, even on a busy org
with tons of users and streams like chat.zulip.org, causes no problem;
so the issue doesn't seem to affect PeopleAutocomplete or
StreamAutocomplete.

Seems like we must have had some pretty severe performance regression
in the emoji autocomplete list.  We should fix that, but first let's
stop the bleeding.  It's currently impossible to scroll around the
list due to another issue (#2917), so we can set the limit tightly
because more emoji than fit on the screen won't do any good anyway.