webview js: Belatedly update generated code that actually runs.
The last few commits that affected `js.js` neglected to update
this file generated from it, which is the version that's actually
used in the app.

Really this just underlines that we ought to improve this pipeline
so the generated version is built automatically, and not checked
into the repo.