webview: Update when a message is starred/unstarred, etc.
Fixes #2676.  When a user stars a message, the long press menu option
"Star message" / "Unstar message" appropriately updates, but the
actual message list UI doesn't.  Since 13c1e6115 we get the flags
information (like `starred`) from the right place when rendering
messages to HTML; but we don't actually re-render the messages when
flags update.  Fix that.

More generally, re-render when any flags change, with one exception:
the `read` flag.  This flag is on the one hand constantly updated
precisely when the user is reading this thread (so it'd be expensive
to constantly re-render); and on the other hand not actually used in
the HTML we render.