narrow: Rename reducer + selector modules from "chat".
In the entire codebase, reducer is named same as its key in the
global state.  So following this style rename `chatReducers` to
`narrowReducers`.  Similarly `chatSelectors` to `narrowSelectors`.

Also "chat" is too generic a word, given that the whole app is about
chat.  So, it's better to rename uses of "chat" to use more
informative terms.

Fixes #2992.