dependencies: Upgrade `react-native-fetch-blob` (as `rn-fetch-blob`)
This fixes a blocker-issue with RN 0.57.2 that prevents the
project from building on Android.

The original project is no longer supported, but a very active
fork is being maintained and named as `rn-fetch-blob`.

This was manually tested and works fine. It contains multiple
fixes for issues that we do not have reports for, but feels nice
to auto-fix issues.

The `notifications` config option is mandatory (as the code does
not handle any default value for it and throws an exception).