package.json: Add `yarn update-libdefs` to update flow-typed/.
This simplifies keeping this directory synced with our changes to our
dependencies, as well as with updates to the upstream flow-typed repo.

The `rm` is to clear out dependencies we've gotten rid of.  There
doesn't seem to be a built-in feature for this in the `flow-typed`
command, leading to this sledgehammer solution; and starting from
scratch this way doesn't make that command materially slower, so
just do it.