webview js: Upgrade generate script to use Babel 7.
This was a bit tricky, because:

 * The old `react-native` preset is incompatible with Babel 7.

 * But the new one is in the module `metro-react-native-babel-preset`...
   and the `module:` syntax for spelling that reference is new in Babel 7.

 * So it doesn't seem to be possible to write a single Babel config
   that works for both versions.  But Jest, and Metro (at least some of
   the time) are still using Babel 6, until we upgrade those too.
   And all of the above look at our .babelrc file.

This babel.config.js, with the `babelrc: false`, is my clever(?)
solution.  It's not even *that* much of a hack, because the new
babel.config.js concept seems to be the recommended way upstream to do
most configuration; after everything's upgraded, I think we'll want to
leave these details in this file, and delete our .babelrc, anyway.

The new Babel version causes a number of small changes in the
generated code, all of them cosmetic: just whitespace, and changing
some single-quotes to double-quotes.